Refactor test mod
