+++ src/testmod/java/nerdhub/cardinal/componentstest/vita/Vita.java
@@ -20,30 +20,18 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package nerdhub.cardinal.componentstest;
+package nerdhub.cardinal.componentstest.vita;
 
 import nerdhub.cardinal.components.api.component.Component;
-import nerdhub.cardinal.components.api.component.ComponentContainer;
-import nerdhub.cardinal.componentstest.vita.EntityVita;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.mob.ZombieEntity;
-import net.minecraft.particle.ParticleTypes;
-import net.minecraft.world.World;
 
-public class VitalityZombieEntity extends ZombieEntity {
-    public VitalityZombieEntity(EntityType<? extends ZombieEntity> type, World world) {
-        super(type, world);
-    }
-
-    @Override
-    public void tickMovement() {
-        super.tickMovement();
-        if (this.world.isClient) {
-            this.world.addParticle(ParticleTypes.DRAGON_BREATH, this.x, this.y + 0.3D, this.z, this.random.nextGaussian() * 0.05D, this.random.nextGaussian() * 0.05D, this.random.nextGaussian() * 0.05D);
-        }
-    }
-
-    protected void initComponents(ComponentContainer<Component> components) {
-        components.put(CardinalComponentsTest.VITA, new EntityVita(this, 20));
+public interface Vita extends Component {
+    int getVitality();
+    void setVitality(int value);
+    default void transferTo(Vita dest, int amount) {
+        int sourceVitality = this.getVitality();
+        int actualAmount = Math.min(sourceVitality, amount);
+        this.setVitality(sourceVitality - actualAmount);
+        dest.setVitality(dest.getVitality() + actualAmount);
     }
 }
+
