+++ src/testmod/resources/assets/componenttest/lang/en_us.json
@@ -1,4 +1,5 @@
 {
   "componenttest:tooltip.vitality": "Vitality stored: %d",
+  "componenttest:tooltip.self_vitality": "Your vitality: %d",
   "item.componenttest.vita_stick": "Vita Stick"
 }
